refactor: printable exceptions and wrapper for formating them
